Authentication code changed

+++ server/src/main/java/com/example/server/service/UserService.java
@@ -1,6 +1,6 @@
 package com.example.server.service;
 
-import com.example.server.model.User;
+import com.example.server.auth.model.user.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
 import java.util.Optional;
@@ -8,8 +8,8 @@ import java.util.Optional;
 public interface UserService extends UserDetailsService {
     User save(User user) throws Exception;
     User patch(User user) throws Exception;
-    User delete(Long id) throws Exception;
-    Optional<User> findById(Long id);
+    User delete(Integer id) throws Exception;
+    Optional<User> findById(Integer id);
     Optional<User> findByEmail(String email);
     Optional<User> findByUsername(String username);
 }
